Fix compiling app_booter

+++ app_booter/source/string.h
@@ -1,9 +1,7 @@
 #ifndef STRING_H_
 #define STRING_H_
 
-#include <stdint.h>
-
-void *memcpy(void *dst, const void *src, size_t len);
-void *memset(void *b, uint8_t c, size_t len);
+void *memcpy(void *dst, const void *src, u32 len);
+void *memset(void *b, u8 c, u32 len);
 
 #endif
